+++ .changeset/config.json
@@ -19,6 +19,8 @@
       "@udecode/slate-plugins-break",
       "@udecode/slate-plugins-code-block",
       "@udecode/slate-plugins-code-block-ui",
+      "@udecode/slate-plugins-color",
+      "@udecode/slate-plugins-color-ui",
       "@udecode/slate-plugins-common",
       "@udecode/slate-plugins-core",
       "@udecode/slate-plugins-csv-serializer",
