+++ package.json
@@ -105,6 +105,7 @@
     "faker": "^5.5.3",
     "jest": "^26.6.3",
     "lerna": "^3.22.1",
+    "node-eval": "^2.0.0",
     "prettier": "^2.2.1",
     "react": "^17.0.1",
     "react-dnd-test-backend": "^12.0.0",
