+++ packages/common/src/transforms/index.ts
@@ -11,8 +11,8 @@ export * from './mergeDeepToNodes';
 export * from './mergeNodes';
 export * from './moveChildren';
 export * from './removeMark';
-export * from './selectEndOfBlockAboveSelection';
 export * from './selectEditor';
+export * from './selectEndOfBlockAboveSelection';
 export * from './setNodes';
 export * from './toggleMark';
 export * from './toggleNodeType';
