+++ packages/dnd/src/index.ts
@@ -2,8 +2,8 @@
  * @file Automatically generated by barrelsby.
  */
 
+export * from './createDndPlugin';
 export * from './types';
 export * from './components/index';
-export * from './createDndPlugin';
 export * from './hooks/index';
 export * from './utils/index';
