+++ packages/marks/color-ui/README.md
@@ -1,0 +1,7 @@
+# slate-plugins-color-ui
+
+See [README](https://github.com/udecode/slate-plugins).
+
+## License
+
+[MIT](../../../LICENSE)
