+++ packages/marks/color-ui/src/ColorPicker/ColorPicker.tsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 
 export const ColorPicker = ({
   color,
