+++ packages/marks/color-ui/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../../scripts/tsconfig.build.json",
+  "compilerOptions": {
+    "declarationDir": "./dist"
+  },
+  "include": ["src"]
+}
