+++ packages/marks/color/.npmignore
@@ -1,0 +1,3 @@
+__tests__
+__test-utils__
+__mocks__
