+++ packages/marks/color/README.md
@@ -1,0 +1,7 @@
+# slate-plugins-color
+
+See [README](https://github.com/udecode/slate-plugins).
+
+## License
+
+[MIT](../../../LICENSE)
