+++ packages/marks/color/src/createColorPlugin.ts
@@ -1,0 +1,9 @@
+import { getRenderLeaf, SlatePlugin } from '@udecode/slate-plugins-core';
+import { MARK_COLOR } from './defaults';
+import { getColorDeserialize } from './getColorDeserialize';
+
+export const createColorPlugin = (): SlatePlugin => ({
+  pluginKeys: MARK_COLOR,
+  renderLeaf: getRenderLeaf(MARK_COLOR),
+  deserialize: getColorDeserialize(),
+});
