+++ packages/marks/color/src/defaults.ts
