+++ packages/marks/color/src/queries/getMark.ts
