+++ packages/serializers/ast-serializer/src/deserializer/index.ts
@@ -1,1 +1,5 @@
+/**
+ * @file Automatically generated by barrelsby.
+ */
+
 export * from './createDeserializeAstPlugin';
