+++ packages/serializers/csv-deserializer/src/deserializer/index.ts
@@ -1,2 +1,6 @@
+/**
+ * @file Automatically generated by barrelsby.
+ */
+
 export * from './createDeserializeCSVPlugin';
 export * from './utils/index';
