+++ packages/serializers/csv-deserializer/src/deserializer/utils/index.ts
@@ -1,1 +1,5 @@
+/**
+ * @file Automatically generated by barrelsby.
+ */
+
 export * from './deserializeCSV';
