+++ packages/serializers/csv-deserializer/src/index.ts
@@ -1,1 +1,5 @@
-export * from './deserializer';
+/**
+ * @file Automatically generated by barrelsby.
+ */
+
+export * from './deserializer/index';
